Share access token extraction for websocket auth
